refactor: addresses -> address 경로 이름 수정
