+++ apps/client/src/components/features/main/MoreList.tsx
@@ -1,0 +1,3 @@
+export default function MoreList() {
+    return <div></div>;
+}
